improved background color selection
Four buttons are now presented to the user in a toggle group: transparent, white, black, and custom. Clicking any of the buttons will switch the background color instantly (as will selecting a custom color in the custom color popup).

The default background color was changed to transparent for clarity. Furthermore, `actingBg` was added (and incorporated into the conversion function) in case the bg was set to transparent when the output format did not support it.