+image file icon background prototype
The list of supported input files have successfully been converted to SVG URI strings in as a proof of concept for creating a dynamic didactic background that subtly informs the users of what file formats are accepted (this also has the upside of potentially filling in the excessive whitespace if an appealing placement pattern can be found).

Thus, because this information has now been more tastefully displayed in an artistic manner, the old checkmark/x mark list has been removed.